API_Youtube/requirements.txt
amqp==5.1.1
asgiref==3.3.1
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==4.1.1
celery==5.2.7
certifi==2020.11.8
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==3.1.4
django-crontab==0.7.1
django-filter==2.4.0
djangorestframework==3.12.2
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
kombu==5.2.4
prompt-toolkit==3.0.38
protobuf==3.15.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.7.1
redis==4.5.1
requests==2.25.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.6